реализовал поиск с помощью спецификаций

+++ main-service/src/main/java/ru/practicum/category/CategoryServiceImpl.java
@@ -11,7 +11,10 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.category.dto.CategoryDto;
 import ru.practicum.category.dto.CategoryShortDto;
 import ru.practicum.category.model.Category;
+import ru.practicum.error.model.ConflictException;
 import ru.practicum.error.model.NotFoundException;
+import ru.practicum.event.model.Event;
+import ru.practicum.event.repository.EventRepository;
 import ru.practicum.util.MapperPageToList;
 
 import java.util.List;
@@ -21,6 +24,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
@@ -32,21 +36,23 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public void deleteCategory(long catId) {
-        log.info("Delete category: {}", catId);
-        categoryRepository.deleteById(catId);
+        Category category = findCategory(catId);
+        List<Event> eventList = eventRepository.findAllByCategory(category);
+        if (eventList.isEmpty()) {
+            categoryRepository.deleteById(catId);
+            log.info("Delete category: {}", catId);
+        } else throw new ConflictException("Category contain Events");
     }
 
     @Override
     @Transactional
     public CategoryDto updateCategory(CategoryShortDto shortDto, long catId) {
-        Category category = categoryRepository.findById(catId).orElseThrow(() ->
-                new NotFoundException(String.format("Category with ID:%d not found", catId)));
+        Category category = findCategory(catId);
         category.setName(shortDto.getName());
 
         log.info("Updating category: {}", category);
         return CategoryMapper.INSTANCE.toDto(category);
     }
-
 
     @Override
     @Transactional(readOnly = true)
@@ -58,10 +64,15 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<CategoryDto> getCategoties(int from, int size) {
+    public List<CategoryDto> getCategories(int from, int size) {
         Sort sortById = Sort.by(Sort.Direction.DESC, "id");
         Pageable page = PageRequest.of(from, size, sortById);
         Page<Category> categoryPage = categoryRepository.findAll(page);
         return MapperPageToList.mapPageToList(categoryPage, from, size, CategoryMapper.INSTANCE::toDto);
+    }
+
+    private Category findCategory(long catId) {
+        return categoryRepository.findById(catId).orElseThrow(() ->
+                new NotFoundException(String.format("Category with ID:%d not found", catId)));
     }
 }
